interface modularized
